# gin-api-server/internal/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "funnydb api doc",
        "title": "FunnyDB",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/apps": {
            "post": {
                "description": "创建应用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用"
                ],
                "summary": "创建应用",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.AppCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppDTO"
                        }
                    }
                }
            }
        },
        "/apps/{id}": {
            "get": {
                "description": "获得应用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用"
                ],
                "summary": "获得应用",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "app id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppDTO"
                        }
                    }
                }
            },
            "put": {
                "description": "更新应用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用"
                ],
                "summary": "更新应用",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "app id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.AppUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除应用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "应用"
                ],
                "summary": "删除应用",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "app id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "{}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/orgs": {
            "post": {
                "description": "创建组织",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "创建组织",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.OrgCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrgDTO"
                        }
                    }
                }
            }
        },
        "/orgs/{id}": {
            "put": {
                "description": "更新组织",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "更新组织",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.OrgUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrgDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除组织",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "删除组织",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "{}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/orgs/{id}/deleted-apps": {
            "get": {
                "description": "获取已删除应用列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "获取已删除应用列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AppDTO"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "恢复已删除应用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "恢复已删除应用",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.OrgRestoreDeletedAppsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/orgs/{id}/members": {
            "get": {
                "description": "获取组织用户列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "获取组织用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "app id",
                        "name": "app_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserDTO"
                            }
                        }
                    }
                }
            }
        },
        "/orgs/{id}/owners": {
            "get": {
                "description": "获取拥有者列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "获取拥有者列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserDTO"
                            }
                        }
                    }
                }
            }
        },
        "/orgs/{id}/roles": {
            "put": {
                "description": "更新组织应用权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "更新组织应用权限",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "app roles",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.OrgUserRoleUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.OrgDTO"
                            }
                        }
                    }
                }
            }
        },
        "/orgs/{id}/users-by-query": {
            "get": {
                "description": "查找用户和应用列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "组织"
                ],
                "summary": "查找用户和应用列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "org id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserWithAppDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AppDTO": {
            "type": "object",
            "properties": {
                "access_key_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "org_id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrgDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "demo@xmfunny.com"
                },
                "email_verified_at": {
                    "type": "integer",
                    "example": 1635696000000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "last_active_at": {
                    "type": "integer",
                    "example": 0
                },
                "logo": {
                    "type": "string",
                    "example": " "
                },
                "name": {
                    "type": "string",
                    "example": "demo"
                }
            }
        },
        "model.UserWithAppDTO": {
            "type": "object",
            "properties": {
                "apps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AppDTO"
                    }
                },
                "email": {
                    "type": "string",
                    "example": "demo@xmfunny.com"
                },
                "email_verified_at": {
                    "type": "integer",
                    "example": 1635696000000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_admin": {
                    "type": "boolean",
                    "example": false
                },
                "is_owner": {
                    "type": "boolean"
                },
                "last_active_at": {
                    "type": "integer",
                    "example": 0
                },
                "logo": {
                    "type": "string",
                    "example": " "
                },
                "name": {
                    "type": "string",
                    "example": "demo"
                }
            }
        },
        "service.AppCreateRequest": {
            "type": "object",
            "required": [
                "identifier",
                "name",
                "org_id"
            ],
            "properties": {
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "org_id": {
                    "type": "integer"
                }
            }
        },
        "service.AppUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.OrgCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "service.OrgRestoreDeletedAppsRequest": {
            "type": "object",
            "required": [
                "app_ids"
            ],
            "properties": {
                "app_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "service.OrgUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.OrgUserRoleUpdateRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "deletes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "email": {
                    "type": "string"
                },
                "orgID": {
                    "type": "integer"
                },
                "updates": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "app_id",
                            "role_id"
                        ],
                        "properties": {
                            "app_id": {
                                "type": "integer"
                            },
                            "role_id": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}